python_ai/requirements.txt
fastapi==0.119.0
uvicorn==0.25.0
numpy<2.0
langchain==0.3.27
langgraph==0.6.10
chromadb==0.4.18
anthropic>=0.16.0
openai>=1.12.0
pydantic>=2.7.4,<3.0.0
sentence-transformers==3.2.0
python-dotenv>=1.0.1
